requirements.txt
ttkbootstrap
pillow
pandas
